+++ cognite/client/utils/_time.py
@@ -1,7 +1,8 @@
 import numbers
 import re
 import time
-from datetime import datetime
+import warnings
+from datetime import datetime, timezone
 from typing import Dict, List, Union
 
 _unit_in_ms_without_week = {"s": 1000, "m": 60000, "h": 3600000, "d": 86400000}
@@ -9,8 +10,15 @@ _unit_in_ms = {**_unit_in_ms_without_week, "w": 604800000}
 
 
 def datetime_to_ms(dt):
-    epoch = datetime.utcfromtimestamp(0)
-    return int((dt - epoch).total_seconds() * 1000.0)
+    if dt.tzinfo is None:
+        warnings.warn(
+            "Interpreting given naive datetime as UTC instead of local time (against Python default behaviour). "
+            "This will change in the next major release (4.0.0). Please use (timezone) aware datetimes "
+            "or convert it yourself to integer (number of milliseconds since epoch, leap seconds excluded).",
+            DeprecationWarning,
+        )
+        dt = dt.replace(tzinfo=timezone.utc)
+    return int(1000 * dt.timestamp())
 
 
 def ms_to_datetime(ms: Union[int, float]) -> datetime:
@@ -20,11 +28,18 @@ def ms_to_datetime(ms: Union[int, float]) -> datetime:
         ms (Union[int, float]): Milliseconds since epoch
 
     Returns:
-        datetime: Datetime object.
+        datetime: Naive datetime object in UTC.
 
     """
     if ms < 0:
         raise ValueError("ms must be greater than or equal to zero.")
+
+    warnings.warn(
+        "This function, `ms_to_datetime` returns a naive datetime object in UTC. This is against "
+        "the default interpretation of naive datetimes in Python (i.e. local time). This behaviour will "
+        "change to returning timezone-aware datetimes in UTC in the next major release (4.0.0).",
+        DeprecationWarning,
+    )
     return datetime.utcfromtimestamp(ms / 1000)